Added boards method
boards.py interacts with boards.json, the 4chan boards list API.
I tested it and it works, and the code should be self explanatory.
However, I did not update the documentation or usage, and the code could
be structured better.
I do not know how to update the documentation properly, so if somebody
else could do so that would be great.